changing to uint256

+++ contracts/interfaces/IVault.sol
@@ -8,17 +8,17 @@ interface IVault is IERC20Metadata {
     error IVault__NotProcessingDeposits();
     error IVault__ForbiddenWhileProcessingDeposits();
 
-    event Deposit(address indexed owner, uint amountDeposited);
-    event Withdraw(address indexed owner, uint sharesBurnt, uint amountWithdrawn);
-    event StartRound(uint indexed roundId, uint amountAddedToStrategy);
-    event EndRound(uint indexed roundId);
-    event DepositProcessed(address indexed owner, uint indexed roundId, uint assets, uint shares);
+    event Deposit(address indexed owner, uint256 amountDeposited);
+    event Withdraw(address indexed owner, uint256 sharesBurnt, uint256 amountWithdrawn);
+    event StartRound(uint256 indexed roundId, uint256 amountAddedToStrategy);
+    event EndRound(uint256 indexed roundId);
+    event DepositProcessed(address indexed owner, uint256 indexed roundId, uint256 assets, uint256 shares);
 
     /**
      * @dev Deposits underlying tokens, generating shares.
      * @param amount The amount of underlying tokens to deposit
      */
-    function deposit(uint amount) external;
+    function deposit(uint256 amount) external;
 
     /**
      * @dev Burn shares, withdrawing underlying tokens.
